refactor: consulta model to entity and move endereçoDTO directory
