+++ src/main/java/com/dh/consultorioOdontologico/entity/Consulta.java
@@ -4,27 +4,21 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
+import javax.persistence.*;
+import java.sql.Timestamp;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@Entity
 public class Consulta {
-    private int id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(nullable = false)
     private int idPaciente;
+    @Column(nullable = false)
     private int idDentista;
-    private LocalDateTime dataConsulta;
-
-    public Consulta(Paciente paciente, Dentista dentista, LocalDateTime dataConsulta) {
-        this.idPaciente = paciente.getId();
-        this.idDentista = dentista.getId();
-        this.dataConsulta = dataConsulta;
-    }
-
-    public Consulta(int idPaciente, int idDentista, LocalDateTime dataConsulta) {
-        this.idPaciente = idPaciente;
-        this.idDentista = idDentista;
-        this.dataConsulta = dataConsulta;
-    }
-
+    @Column(nullable = false)
+    private Timestamp dataConsulta;
 }
